# backend/swagger.yml
openapi: 3.0.3
info:
  title: Kids Jobs Backend API
  description: |
    API standalone do kids-jobs para vagas, fontes, scrapings
    e operacao local.
  version: 2.0.0
servers:
  - url: http://localhost:8001
    description: Local development
...
paths:
...
  /api/v1/sources:
    get:
      tags: [sources]
      summary: List all sources
      operationId: listSources
      parameters:
        - name: enabled_only
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Source" }
    post:
      tags: [sources]
      summary: Create source
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "409":
          description: Conflict (duplicate name)
  /api/v1/sources/{source_id}:
    get:
      tags: [sources]
      summary: Get source by ID
      operationId: getSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Source details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404":
          description: Not found
    put:
      tags: [sources]
      summary: Update source
      operationId: updateSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404":
          description: Not found
    delete:
      tags: [sources]
      summary: Delete source
      operationId: deleteSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /api/v1/sources/{source_id}/toggle:
    patch:
      tags: [sources]
      summary: Toggle source enabled/disabled
      operationId: toggleSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404":
          description: Not found
  /api/v1/sources/{source_id}/analysis:
    patch:
      tags: [sources]
      summary: Update source analysis
      operationId: updateSourceAnalysis
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceAnalysisUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404":
          description: Not found
  /api/v1/source-executions:
    get:
      tags: [source-executions]
      summary: List source execution history
      operationId: listSourceExecutions
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [24h, 7d, 30d, 90d]
            default: 30d
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 5000
      responses:
        "200":
          description: Recent source execution history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SourceExecutionHistory" }

  # --- Categories -----------------------------------------------------------
...
components:
  schemas:
...
    Source:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        enabled: { type: boolean }
        scraper_base_url: { type: string }
        scraper_type: { type: string }
        scraper_schedule: { type: string }
        extra_config: { type: object, additionalProperties: true }
        analysis: { type: string }
        description: { type: string }
        last_extraction_status: { type: string, nullable: true, example: SUCCESS }
        last_extraction_http_status: { type: integer, nullable: true, example: 200 }
        last_extraction_message: { type: string, nullable: true }
        last_extraction_at: { type: string, format: date-time, nullable: true }
        next_scheduled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SourceCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        enabled: { type: boolean, default: true }
        scraper_base_url: { type: string }
        scraper_type: { type: string, default: http }
        scraper_schedule: { type: string, default: "0 */2 * * *" }
        extra_config: { type: object, additionalProperties: true }
        analysis: { type: string, nullable: true }
        description: { type: string, nullable: true }
    SourceUpdate:
      type: object
      properties:
        name: { type: string }
        enabled: { type: boolean }
        scraper_base_url: { type: string }
        scraper_type: { type: string }
        scraper_schedule: { type: string }
        extra_config: { type: object, additionalProperties: true }
        analysis: { type: string, nullable: true }
        description: { type: string, nullable: true }
    SourceExecutionHistory:
      type: object
      properties:
        id: { type: integer }
        source_id: { type: integer }
        source_name: { type: string }
        trigger: { type: string, example: scheduled }
        status: { type: string, example: SUCCESS }
        success: { type: boolean }
        scraped_count: { type: integer }
        published_count: { type: integer }
        duration_ms: { type: integer }
        strategy: { type: string, example: http_antibot }
        http_status_code: { type: integer, nullable: true }
        error_message: { type: string, nullable: true }
        message: { type: string, nullable: true }
        executed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
    SourceAnalysisUpdate:
      type: object
      required: [analysis]
      properties:
        analysis: { type: string }
...